refactor: 관심사에 따른Layer 재분배
