+++ src/Models/Fetcher.js
@@ -1,8 +1,13 @@
 import { ERROR } from '../constants';
 
-export const Fetcher = {
+export class Fetcher {
+  #data = null;
+  #isLoading = false;
+
   async get(endpoint, config = {}) {
     const { method, headers } = config;
+    this.#isLoading = true;
+
     const response = await fetch(endpoint, {
       method,
       headers,
@@ -10,6 +15,8 @@ export const Fetcher = {
 
     if (!response.ok) throw new Error(ERROR.HTTP(response.status));
 
+    this.#isLoading = false;
+
     return await response.json();
-  },
-};
+  }
+}
