+++ src/Models/Movie/MovieController.js
@@ -1,27 +1,12 @@
 import { MovieService } from './MovieService';
-import { MOVIE_API, PROMISE_STATE } from '../../constants';
+import { MOVIE_API } from '../../constants';
+
+const movieApis = [MOVIE_API.TMDB.POPULAR_MOVIE];
 
 export class MovieController {
-  #service = new MovieService();
+  #service = new MovieService(movieApis);
 
-  async fetchAllMovies() {
-    const movieApiConfigs = [MOVIE_API.TMDB.POPULAR_MOVIE];
-    const movieResults = await this.#fetchMoviesFromApis(movieApiConfigs);
-
-    return this.#extractFulfilledResults(movieResults);
-  }
-
-  async #fetchMoviesFromApis(apiConfigs) {
-    return await Promise.allSettled(
-      apiConfigs.map(({ endpoint, config }) =>
-        this.#service.getMovies(endpoint, config)
-      )
-    );
-  }
-
-  #extractFulfilledResults(results) {
-    return results
-      .filter(({ status }) => status === PROMISE_STATE.FULFILLED)
-      .map(({ value }) => value);
+  getAllMovies() {
+    return this.#service.fetchAllMovies();
   }
 }
